test(): added testing with delete category

+++ src/modules/category/service/category.service.ts
@@ -3,12 +3,16 @@ import { CreateCategoryDto, UpdateCategoryDto } from '../dto/category.dto';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { Category } from '../schema/category.schema';
+import { Product } from '../../product/schema/product.schema';
 
 @Injectable()
 export class CategoryService {
 
     // GET COLLECTIONS INSTANCE
-    constructor(@InjectModel(Category.name) private categoryModel: Model<Category>) { }
+    constructor(
+        @InjectModel(Category.name) private categoryModel: Model<Category>,
+        @InjectModel(Product.name) private productModel: Model<Product>
+    ) { }
 
     /**
      * Find all categories
@@ -69,6 +73,17 @@ export class CategoryService {
      * @param id
      */
     async delete(id: string) {
+
+        // CHECK IF CATEGORY HAS USED BY ANY PRODUCT
+        const isUsed = await this.productModel.findOne({ categoryId: id });
+        if (isUsed) {
+            return {
+                message: 'Category is used by some product, cannot delete',
+                success: false,
+                data: null
+            }
+        }
+
         await this.categoryModel.findByIdAndDelete(id);
 
         return {
